requirements.txt
###### Requirements without Version Specifiers ######
numpy
pandas
plotly
streamlit
google-cloud-firestore
PyDrive